Move things around and add license headers
