+++ external-crates/move/crates/move-package-alt/src/lib.rs
@@ -1,4 +1,5 @@
-// Copyright (c) Mysten Labs, Inc.
+// Copyright (c) The Diem Core Contributors
+// Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
 //! This library defines the package management system for Move.
