+++ external-crates/move/crates/move-package-alt/src/package/lockfile.rs
@@ -1,3 +1,7 @@
+// Copyright (c) The Diem Core Contributors
+// Copyright (c) The Move Contributors
+// SPDX-License-Identifier: Apache-2.0
+
 use derivative::Derivative;
 use lazy_regex::regex_captures;
 use serde_spanned::Spanned;
